move training-module to a subfolder
