+++ src/tutorial-module/training-module/WelcomeTraining.tsx
@@ -1,12 +1,12 @@
 import React from "react";
 import styled from "styled-components";
 
-import { MainButton } from "../webapp/components/main-button/MainButton";
-import { Modal, ModalContent, ModalFooter } from "../webapp/components/modal";
-import { MarkdownViewer } from "../webapp/components/markdown-viewer/MarkdownViewer";
-import i18n from "../utils/i18n";
-import { TrainingModule } from "../domain/entities/TrainingModule";
-import { TranslateMethod } from "../domain/entities/TranslatableText";
+import { MainButton } from "../../webapp/components/main-button/MainButton";
+import { Modal, ModalContent, ModalFooter } from "../../webapp/components/modal";
+import { MarkdownViewer } from "../../webapp/components/markdown-viewer/MarkdownViewer";
+import i18n from "../../utils/i18n";
+import { TrainingModule } from "../../domain/entities/TrainingModule";
+import { TranslateMethod } from "../../domain/entities/TranslatableText";
 import { HeaderButtonsProps } from "./TutorialRoot";
 
 type WelcomeTrainingProps = {
